reflects comments on the PR
